Add update types to fields in JSON schemas (WIP)
Define for each field whether it is updated manually ('manual') has to
be freely researched ('research') or just can be fetched from a known
source ('fetch').

Note that this is temporary work and the design moves into a different
direction where this logic is contained in the agent that does the
updates.